# plan: calc `access path` when doing `deriveStats`.