--- a/plan/build_key_info.go
+++ b/plan/build_key_info.go
@@ -174,8 +174,11 @@ func (p *LogicalJoin) buildKeyInfo() {
 func (ds *DataSource) buildKeyInfo() {
 	ds.schema.Keys = nil
 	ds.baseLogicalPlan.buildKeyInfo()
-	indices := ds.availableIndices.indices
-	for _, idx := range indices {
+	for _, path := range ds.possibleIndexPaths {
+		if path.isRowID {
+			continue
+		}
+		idx := path.index
 		if !idx.Unique {
 			continue
 		}