--- a/plan/build_key_info.go
+++ b/plan/build_key_info.go
@@ -175,7 +175,7 @@ func (ds *DataSource) buildKeyInfo() {
 	ds.schema.Keys = nil
 	ds.baseLogicalPlan.buildKeyInfo()
 	for _, path := range ds.possibleAccessPaths {
-		if path.isRowID {
+		if path.isTablePath {
 			continue
 		}
 		idx := path.index