… plan/plan.go
… plan/planbuilder.go
… plan/property_cols_prune.go
… plan/stats.go