- collect thing as a struct.
- reduce redundant calculation.
- don't rely on `statsInfo`.