@winoros 2018-04-23
simple benchmark result
previous:
```
BenchmarkOptimize/select_count(*)_from_t_group_by_e-8         	   30000	     47771 ns/op	   38058 B/op	     500 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_10_group_by_e-8         	   20000	    106329 ns/op	   51307 B/op	     730 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_50-8                    	   20000	     73301 ns/op	   36746 B/op	     561 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_c_>_'1'_group_by_b-8         	   10000	    104546 ns/op	   55571 B/op	     811 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_=_1_group_by_b-8           	   10000	    103423 ns/op	   59859 B/op	     871 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_>_1_group_by_b-8           	   10000	    117617 ns/op	   62900 B/op	     888 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_20-8                  	   10000	    101968 ns/op	   60443 B/op	     848 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_30-8                  	   10000	    112738 ns/op	   60443 B/op	     848 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_40-8                  	   10000	    102269 ns/op	   60443 B/op	     848 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_50-8                  	   10000	    103319 ns/op	   60443 B/op	     848 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_40-8                         	   30000	     47411 ns/op	   27488 B/op	     359 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_50-8                         	   30000	     49996 ns/op	   27488 B/op	     359 allocs/op
BenchmarkOptimize/select_*_from_t_where_1_and_t.b_<=_50-8                   	   30000	     46535 ns/op	   27696 B/op	     364 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_100_order_by_t.a_limit_1-8   	   10000	    191054 ns/op	  119193 B/op	    1527 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_1_order_by_t.a_limit_10-8    	   10000	    181012 ns/op	  119192 B/op	    1527 allocs/op
BenchmarkOptimize/select_*_from_t_use_index(b)_where_b_=_1_order_by_a-8     	   30000	     39580 ns/op	   24854 B/op	     350 allocs/op
BenchmarkOptimize/select_*_from_t_where_d_<_cast('1991-09-05'_as_datetime)-8         	   30000	     42636 ns/op	   21480 B/op	     296 allocs/op
plan: calc `index path` when doing `deriveStats`.
BenchmarkOptimize/select_*_from_t_where_ts_<_'1991-09-05'-8                          	   30000	     44727 ns/op	   21384 B/op	     294 allocs/op
```

Now:
```
BenchmarkOptimize/select_count(*)_from_t_group_by_e-8         	   30000	     46541 ns/op	   39234 B/op	     495 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_10_group_by_e-8         	   20000	     68654 ns/op	   47858 B/op	     645 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_50-8                    	   30000	     55205 ns/op	   33458 B/op	     492 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_c_>_'1'_group_by_b-8         	   20000	     83212 ns/op	   54291 B/op	     744 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_=_1_group_by_b-8           	   20000	     80127 ns/op	   57179 B/op	     750 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_>_1_group_by_b-8           	   20000	     90272 ns/op	   58171 B/op	     755 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_20-8                  	   20000	     80168 ns/op	   52451 B/op	     693 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_30-8                  	   20000	     77870 ns/op	   52451 B/op	     693 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_40-8                  	   20000	     79757 ns/op	   52451 B/op	     693 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_50-8                  	   20000	     80678 ns/op	   52450 B/op	     693 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_40-8                         	   30000	     52173 ns/op	   30704 B/op	     447 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_50-8                         	   30000	     54103 ns/op	   30704 B/op	     447 allocs/op
BenchmarkOptimize/select_*_from_t_where_1_and_t.b_<=_50-8                   	   30000	     53793 ns/op	   30912 B/op	     452 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_100_order_by_t.a_limit_1-8   	   10000	    111318 ns/op	   98032 B/op	    1033 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_1_order_by_t.a_limit_10-8    	   10000	    115871 ns/op	   98032 B/op	    1033 allocs/op
BenchmarkOptimize/select_*_from_t_use_index(b)_where_b_=_1_order_by_a-8     	   30000	     38680 ns/op	   25318 B/op	     338 allocs/op
BenchmarkOptimize/select_*_from_t_where_d_<_cast('1991-09-05'_as_datetime)-8         	   30000	     52790 ns/op	   25760 B/op	     410 allocs/op
BenchmarkOptimize/select_*_from_t_where_ts_<_'1991-09-05'-8                          	   30000	     58391 ns/op	   25664 B/op	     408 allocs/op
```


If the sql have `prop`s to push down, it will reduce a lot of time.
But if there's no `prop`, it will cost more because all index calculation is done.
We can use a lazy calcultion to optimize this.