@winoros 2018-04-23
sysbench
select: 2% improve.
oltp: not changed too much.