@winoros 2018-04-23
/run-all-tests