@winoros 2018-04-25
/run-all-tests